+++ p2panda-rs/src/test_utils/mocks/client.rs
@@ -14,13 +14,11 @@
 //!
 //! use p2panda_rs::operation::AsOperation;
 //! use p2panda_rs::operation::{OperationEncoded, OperationValue};
-//! use p2panda_rs::test_utils::constants::DEFAULT_SCHEMA_HASH;
+//! use p2panda_rs::test_utils::constants::TEST_SCHEMA_ID;
 //! use p2panda_rs::test_utils::mocks::{send_to_node, Client, Node};
 //! use p2panda_rs::test_utils::utils::{
 //!     create_operation, schema, new_key_pair, operation_fields, update_operation,
 //! };
-//!
-//! # const CHAT_SCHEMA_HASH: &str = DEFAULT_SCHEMA_HASH;
 //!
 //! // Instantiate a new mock node
 //! let mut node = Node::new();
@@ -30,7 +28,7 @@
 //!
 //! // Create a new operation to publish
 //! let operation = create_operation(
-//!     schema(DEFAULT_SCHEMA_HASH),
+//!     schema(TEST_SCHEMA_ID),
 //!     operation_fields(vec![(
 //!         "message",
 //!         OperationValue::Text("Ohh, my first message!".to_string()),