+++ p2panda-rs/src/test_utils/utils.rs
@@ -16,7 +16,7 @@ use crate::operation::{
     Operation, OperationEncoded, OperationFields, OperationId, OperationValue, OperationWithMeta,
 };
 use crate::schema::SchemaId;
-use crate::test_utils::constants::DEFAULT_SCHEMA_HASH;
+use crate::test_utils::constants::TEST_SCHEMA_ID;
 
 /// A custom `Result` type to be able to dynamically propagate `Error` types.
 pub type Result<T> = std::result::Result<T, Box<dyn std::error::Error>>;
@@ -46,24 +46,18 @@ pub fn any_operation(
     fields: Option<OperationFields>,
     previous_operations: Option<Vec<OperationId>>,
 ) -> Operation {
+    let schema_id = SchemaId::new(TEST_SCHEMA_ID).unwrap();
     match fields {
         // It's a CREATE operation
         Some(fields) if previous_operations.is_none() => {
-            Operation::new_create(SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap(), fields).unwrap()
+            Operation::new_create(schema_id, fields).unwrap()
         }
         // It's an UPDATE operation
-        Some(fields) => Operation::new_update(
-            SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap(),
-            previous_operations.unwrap(),
-            fields,
-        )
-        .unwrap(),
+        Some(fields) => {
+            Operation::new_update(schema_id, previous_operations.unwrap(), fields).unwrap()
+        }
         // It's a DELETE operation
-        None => Operation::new_delete(
-            SchemaId::new(DEFAULT_SCHEMA_HASH).unwrap(),
-            previous_operations.unwrap(),
-        )
-        .unwrap(),
+        None => Operation::new_delete(schema_id, previous_operations.unwrap()).unwrap(),
     }
 }
 
@@ -96,9 +90,9 @@ pub fn hash(hash_str: &str) -> Hash {
     Hash::new(hash_str).unwrap()
 }
 
-/// Generate an application schema based on a hash string.
-pub fn schema(hash_str: &str) -> SchemaId {
-    SchemaId::new(hash_str).unwrap()
+/// Generate an application schema based on a schema id string.
+pub fn schema(schema_id: &str) -> SchemaId {
+    SchemaId::new(schema_id).unwrap()
 }
 
 /// Generate an entry based on passed values.
@@ -143,23 +137,4 @@ pub fn meta_operation(
     operation_encoded: OperationEncoded,
 ) -> OperationWithMeta {
     OperationWithMeta::new(&entry_signed_encoded, &operation_encoded).unwrap()
-}
-
-#[cfg(test)]
-mod tests {
-    use crate::test_utils::constants::DEFAULT_HASH;
-
-    use super::*;
-
-    #[test]
-    fn default_hash() {
-        let default_hash = Hash::new_from_bytes(vec![1, 2, 3]).unwrap();
-        assert_eq!(default_hash.as_str(), DEFAULT_HASH)
-    }
-
-    #[test]
-    fn default_schema() {
-        let default_schema_hash = Hash::new_from_bytes(vec![3, 2, 1]).unwrap();
-        assert_eq!(default_schema_hash.as_str(), DEFAULT_SCHEMA_HASH)
-    }
 }